+++ test/controllers/settings_controller_test.rb
@@ -20,24 +20,15 @@ class SettingsControllerTest < ActionDispatch::IntegrationTest
     sign_in_as(user)
     get settings_url
     assert_response :success
-    assert_select "[data-key='settings.email']", text: /Your current email is #{Regexp.escape(user.email_address)}/
-    assert_select "[data-key='settings.password']", text: /Last changed .+ ago/
+    assert_select "[data-key='settings.email'] p", text: "Your current email is #{user.email_address}"
+    assert_select "[data-key='settings.password'] p", text: /\ALast changed .+ ago\z/
   end
 
-  test "should show the current name in the account card" do
-    @user = create(:user, name: "Alex")
+  test "should describe the name card" do
     sign_in_as(user)
     get settings_url
     assert_response :success
-    assert_select "[data-key='settings.name']", text: /People you invite will see Alex/
-  end
-
-  test "should explain the fallback when no name is set" do
-    @user = create(:user, name: "")
-    sign_in_as(user)
-    get settings_url
-    assert_response :success
-    assert_select "[data-key='settings.name']", text: /Somebody/
+    assert_select "[data-key='settings.name'] p", text: "The name people you invite will see"
   end
 
   test "should link to settings sections" do